requirements.txt
absl-py==2.1.0
asttokens 
astunparse==1.6.3
backcall 
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
comm 
contourpy==1.1.1
cycler==0.12.1
debugpy 
decorator 
executing 
filelock==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.0
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.64.1
gunicorn==22.0.0
h5py==3.11.0
idna==3.7
imageio==2.34.1
importlib_metadata 
importlib_resources==6.4.0
ipykernel 
ipython 
itsdangerous==2.2.0
jedi 
Jinja2==3.1.4
joblib==1.4.2
jupyter_client 
jupyter_core 
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline 
mpmath==1.3.0
nest_asyncio 
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu12 
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12 
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.10.0.82
opt-einsum==3.3.0
packaging==24.0
pandas==2.0.3
parso 
pexpect 
pickleshare 
pillow==10.3.0
platformdirs 
prompt_toolkit 
protobuf==4.25.3
psutil 
ptyprocess 
pure-eval 
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyglet==2.0.15
Pygments 
pyparsing==3.1.2
python-dateutil 
pytz==2024.1
PyYAML==6.0.1
pyzmq 
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six 
stack-data 
sympy==1.12.1
termcolor==2.4.0
Theano==1.0.5
threadpoolctl==3.5.0
torch==2.3.1
torchvision==0.18.1
tornado 
tqdm==4.66.4
traitlets 
triton==2.3.1
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
wcwidth 
Werkzeug==3.0.3
wrapt==1.16.0
zipp 
